Move shared mocks to test/shared/mocks.dart
