+++ app/test/frontend/handlers_test.dart
@@ -19,6 +19,7 @@ import 'package:pub_dartlang_org/shared/analyzer_client.dart';
 import 'package:pub_dartlang_org/shared/search_service.dart';
 
 import '../shared/handlers_test_utils.dart';
+import '../shared/mocks.dart';
 import '../shared/utils.dart';
 
 import 'handlers_test_utils.dart';
@@ -221,7 +222,7 @@ void main() {
             1,
             [
               new PackageVersionView(
-                  testPackageVersion, new MockAnalysisView()),
+                  testPackageVersion, new AnalysisViewMock()),
             ],
           );
         }));
@@ -239,7 +240,7 @@ void main() {
             1,
             [
               new PackageVersionView(
-                  testPackageVersion, new MockAnalysisView()),
+                  testPackageVersion, new AnalysisViewMock()),
             ],
           );
         }));
