+++ app/test/shared/utils.dart
@@ -7,8 +7,6 @@ import 'dart:async';
 import 'package:gcloud/service_scope.dart';
 import 'package:test/test.dart';
 
-import 'package:pub_dartlang_org/shared/analyzer_client.dart';
-
 Future scoped(func()) {
   return fork(() async {
     return func();
@@ -21,29 +19,4 @@ void scopedTest(String name, func(), {Timeout timeout}) {
       return func();
     });
   }, timeout: timeout);
-}
-
-class MockAnalysisView implements AnalysisView {
-  @override
-  bool hasAnalysisData = true;
-
-  @override
-  AnalysisStatus analysisStatus;
-
-  @override
-  List<String> getTransitiveDependencies() => throw 'Not implemented';
-
-  @override
-  double health;
-
-  @override
-  String licenseText;
-
-  @override
-  DateTime timestamp;
-
-  @override
-  List<String> platforms;
-
-  MockAnalysisView({this.platforms});
 }
